fix: LSP uinteger bound, nested yields, walrus kwargs, honest CLI errors
The last-line merge edit ends at i32::MAX instead of u32::MAX — the
LSP uinteger cap is 2^31-1 and JVM clients parse character as a signed
int. find_yield_offset recurses into expression containers so nested
yields (x = f((yield 1))) mark generators, while lambda yields stay
scoped to the lambda. collect_walrus_targets covers keyword arguments
and callees. The fixtures-unused JSON path reports serialization errors
on stderr instead of printing an empty array that reads as success.

+++ src/providers/code_action.rs
@@ -148,7 +148,10 @@ fn emit_kind_import_edits(
                     (
                         Position {
                             line: fi.end_line as u32,
-                            character: u32::MAX,
+                            // LSP uinteger max (2^31 - 1): past any real line
+                            // length, so clients clamp to end-of-line, and it
+                            // fits the signed 32-bit ints JVM clients use.
+                            character: i32::MAX as u32,
                         },
                         merged_line.clone(),
                     )
@@ -1085,7 +1088,7 @@ mod tests {
         assert_eq!(edits.len(), 1);
         assert_eq!(edits[0].range.start.line, 0);
         assert_eq!(edits[0].range.end.line, 0);
-        assert_eq!(edits[0].range.end.character, u32::MAX);
+        assert_eq!(edits[0].range.end.character, i32::MAX as u32);
         assert_eq!(edits[0].new_text, "from typing import Any, Optional");
     }
